Fix all pop using a single BitArray for genotype
Fix a bug where a selfish, parasitic genotype attached itself in a
reference to every individual it came across like a virus.
Basically the entire population had a single BitArray as a shared
genome which, while annoying, was quite a cool example of a genetic
disease moving through a population thanks to asexual reproduction.